refactor: 카드 유효성 검사 모듈 도메인 레이어로 이동
